Refactors cli.py into different files
