+++ brainsets/_cli/cli_list.py
@@ -1,0 +1,11 @@
+import click
+
+from .utils import DATASETS
+
+
+@click.command()
+def cli_list():
+    """List available datasets."""
+    click.echo("Available datasets:")
+    for dataset in DATASETS:
+        click.echo(f"- {dataset}")
